+++ .gitignore
@@ -9,12 +9,16 @@
 /test/tmp/
 /test/version_tmp/
 /tmp/
+.envrc
+Gemfile.lock
 
 # Used by dotenv library to load environment variables.
 # .env
 
 # Ignore Byebug command history file.
 .byebug_history
+.ruby-version
+.DS_Store
 
 ## Specific to RubyMotion:
 .dat*
