+++ lib/shop_schema_client/shop_schema_composer/metaobject_definition.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+module ShopSchemaClient
+  class ShopSchemaComposer
+    MetaobjectDefinition = Struct.new(
+      :id,
+      :type,
+      :description,
+      :fields,
+      keyword_init: true
+    ) do
+      class << self
+        def from_graphql(metaobject_def)
+          new(
+            id: metaobject_def["id"],
+            type: metaobject_def["type"],
+            description: metaobject_def["description"],
+            fields: metaobject_def["fieldDefinitions"].map { MetafieldDefinition.from_graphql(_1) },
+          )
+        end
+      end
+    end
+  end
+end
