+++ lib/shop_schema_client/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ShopSchemaClient
+  VERSION = "0.0.1"
+end
