+++ test/fixtures/admin_2025_01_public.graphql
