+++ test/fixtures/shop_metaschema.graphql
