+++ test/fixtures/shop_metaschema.json
